Separate CLI and bot's API and controller.
