+++ timeflies/utils.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 #   encoding: utf8
 #   utils.py
 
@@ -10,15 +9,16 @@ from enum import Enum
 from pprint import pprint
 from requests import Session
 
-from .settings import *
-from .models import User, Timelapse
+from .settings import API_URL, API_TOKEN
+from .models import User
 
 
 class Command(Enum):
     # TODO add other commands
     start = 'start'
 
-command_patterns = {c:re.compile('/%s' % c.name) for c in Command}
+
+command_patterns = {c: re.compile('/%s' % c.name) for c in Command}
 
 
 def send_request(method=None, params=None, sess=None):
@@ -27,7 +27,7 @@ def send_request(method=None, params=None, sess=None):
 
     url = API_URL.format(token=API_TOKEN, method=method)
     r = sess.get(url, params=params)
-    
+
     if r.status_code != 200:
         logging.error('request failed with status code %d', r.status_code)
         return None
@@ -41,10 +41,15 @@ def send_request(method=None, params=None, sess=None):
     try:
         json = r.json()
     except ValueError:
-        logging.error('invalid json: %s', res.text)
+        logging.error('invalid json: %s', r.text)
         return None
-    
+
     return json
+
+
+def get_updates(offset=None, limit=100, timeout=60, sess=None):
+    params = dict(offset=offset, limit=limit, timeout=60)
+    return send_request('getUpdates', params, sess)
 
 
 def send_message(chat_id, text, sess=None):
@@ -52,16 +57,16 @@ def send_message(chat_id, text, sess=None):
     return send_request('sendMessage', params, sess)
 
 
-def add_user(user_info, database): 
+def add_user(user_info, database):
     try:
         telegram_id = user_info.get('id')
         query = database.query(User).filter(User.id == telegram_id)
         if query.count():
             raise Exception('User already in database (id %d)' % telegram_id)
-        
+
         username = user_info.get('username')
         first_name = user_info.get('first_name')
-        
+
         user = User(username=username, id=telegram_id, first_name=first_name)
         database.add(user)
         database.commit()
@@ -71,19 +76,19 @@ def add_user(user_info, database):
 
 
 def detect_commands(message):
-    # TODO should parse all command types and return a list of commands (in case
-    # message contains more than one). 
+    # TODO should parse all command types and return a list of commands (in
+    # case message contains more than one).
 
     f = False
     commands = []
-    entities = message.get('entities',[])
+    entities = message.get('entities', [])
     text = message.get('text')
     for e in entities:
         if e.get('type') == 'bot_command':
             f = True
             break
     if f:
-        for c,p in command_patterns.items():
+        for c, p in command_patterns.items():
             if p.search(text):
                 commands.append(c.name)
     logging.info('received commands: %s' % commands)
@@ -92,10 +97,10 @@ def detect_commands(message):
 
 def handle_commands(commands, message, database):
     # TODO handle other commands
-    
+
     user_info = message.get('from')
     for c in commands:
-        if c==Command.start.name:
+        if c == Command.start.name:
             add_user(user_info, database)
 
 
@@ -113,12 +118,12 @@ def handle_update(update, session, database):
 
 def echo(update, sess):
     m = update.get('message')
-    
+
     if m:
         chat_id = m.get('chat').get('id')
         username = m.get('from').get('username')
         message_text = m.get('text')
-        
+
         text = '%s said %s' % (username, message_text)
         send_message(chat_id, text, sess)
         logging.info(update.get('update_id'))
